if we cannot modify bright we don't show on i3blocks

+++ I3/.bin/backlight.sh
@@ -1,3 +1,7 @@
 #!/usr/bin/env bash
-echo "$(xbacklight $@ | cut -d '.' -f 1) "
-pkill -RTMIN+2 i3blocks
+xbacklight $@
+if [ $? -eq 0  ]
+then
+    echo "$(xbacklight $@ | cut -d '.' -f 1) "
+    pkill -RTMIN+2 i3blocks
+fi
